+++ src/main/java/edu/mondragon/webengl/domain/encuesta/repository/EncuestaRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import edu.mondragon.webengl.domain.encuesta.model.Encuesta;
 @Repository
-public interface EncuestaRepository extends JpaRepository<Encuesta, Short> {
+public interface EncuestaRepository extends JpaRepository<Encuesta, Integer> {
+
+
 }
