+++ src/main/java/edu/mondragon/webengl/domain/evento/repository/EventoLocalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import edu.mondragon.webengl.domain.evento.model.EventoLocal;
 
 @Repository
-public interface EventoLocalRepository extends JpaRepository<EventoLocal, Short> {
+public interface EventoLocalRepository extends JpaRepository<EventoLocal, Integer> {
 }
