+++ src/main/java/edu/mondragon/webengl/domain/user/repository/UsuarioRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import edu.mondragon.webengl.domain.user.model.Usuario;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Short> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
     Optional<Usuario> findByEmail(String email);
     //El optional es una clase en el que el contenido puede contener un usuario(en este caso) o pude estar vacío.
     Optional<Usuario> findByUsername(String username);
